feat: Decouple parsing from forwarding using `TeeForwardConn` to minimize recording latency and refactor TCP socket options into a new orchestrator package.
