+++ pkg/agent/proxy/orchestrator/tee_forward_conn.go
@@ -8,7 +8,6 @@ import (
 	"net"
 	"sync"
 	"sync/atomic"
-	"syscall"
 	"time"
 
 	"go.uber.org/zap"
@@ -16,10 +15,6 @@ import (
 
 // Ensure ringSignal is used (defined in ring_signal_linux.go / ring_signal_others.go).
 var _ = (*ringSignal)(nil)
-
-// TCP_QUICKACK is the Linux socket option to disable delayed ACKs.
-// This constant is not in the standard syscall package for all platforms.
-const TCP_QUICKACK = 12
 
 // ringBuf is a lock-free single-producer / single-consumer (SPSC) ring buffer
 // optimised for the TeeForwardConn use-case.
@@ -425,26 +420,6 @@ func SetTCPNoDelay(conn net.Conn) {
 // setTCPNoDelay is the package-internal alias (used by constructors).
 func setTCPNoDelay(conn net.Conn) { SetTCPNoDelay(conn) }
 
-// SetTCPQuickACK disables delayed ACKs to reduce read-side latency.
-// Delayed ACKs can add up to 40ms on Linux when the proxy reads from
-// one connection and writes to another (the piggyback ACK path never fires).
-// NOTE: Linux resets TCP_QUICKACK after every ACK, so this must be called
-// repeatedly — ideally after every Write() that precedes a Read().
-// Handles TLS-wrapped connections by unwrapping to the underlying *net.TCPConn.
-func SetTCPQuickACK(conn net.Conn) {
-	tc := unwrapTCPConn(conn)
-	if tc == nil {
-		return
-	}
-	rawConn, err := tc.SyscallConn()
-	if err != nil {
-		return
-	}
-	_ = rawConn.Control(func(fd uintptr) {
-		_ = syscall.SetsockoptInt(int(fd), syscall.IPPROTO_TCP, TCP_QUICKACK, 1)
-	})
-}
-
 // setTCPQuickACK is the package-internal alias (used by constructors).
 func setTCPQuickACK(conn net.Conn) { SetTCPQuickACK(conn) }
 
@@ -468,36 +443,6 @@ func unwrapTCPConn(conn net.Conn) *net.TCPConn {
 			return nil
 		}
 	}
-}
-
-// extractTCPfd extracts and caches the file descriptor from a TCP connection.
-// Returns -1 if the connection is not TCP. This avoids calling SyscallConn()
-// on every forwarded packet in the hot path.
-// Handles TLS-wrapped connections by unwrapping to the underlying *net.TCPConn.
-func extractTCPfd(conn net.Conn) int {
-	tc := unwrapTCPConn(conn)
-	if tc == nil {
-		return -1
-	}
-	rawConn, err := tc.SyscallConn()
-	if err != nil {
-		return -1
-	}
-	fd := -1
-	_ = rawConn.Control(func(f uintptr) {
-		fd = int(f)
-	})
-	return fd
-}
-
-// quickACKByFD re-enables TCP_QUICKACK using a cached file descriptor.
-// This is the hot-path version — avoids SyscallConn() + Control() overhead
-// (~2-5μs saved per call, significant at thousands of packets/sec).
-func quickACKByFD(fd int) {
-	if fd < 0 {
-		return
-	}
-	_ = syscall.SetsockoptInt(fd, syscall.IPPROTO_TCP, TCP_QUICKACK, 1)
 }
 
 // Read returns buffered data that was already forwarded to dest.
